Menu enchancments and search improvments
The menu has been updated some so it is easier to read and follow. It also allows for the user to enter more options. Showing items through search looks a whole lot better now. The user can still crash the program if they go to far to the right or left in the entries, but that will be fixed later on.