Work Log V1.0
The work log application is up and running and is functional with add and search for features. Search shows you a slideshow more or less that you can go through. I still need to add a way to delete and edit entries. Searching via a range of dates instead of just one needs to be added as well.